requirements.txt
pillow==10.3.0
requests==2.32.3
pyTelegramBotAPI==4.20.0
python-dotenv==1.1.1